gdb/lkd: Ensure async operations are disabled
The LKD implementation does not support asyncronous targets, and expects
'to_wait' to block. LKD assumes that when to_wait completes, it can reload
the threading information. On an async target such as QEmu, this causes
an attempt to read memory from a running target and fails. Leaving the remote
in an unreliable state