gdb/command: Add LKD Command class
LKD's commands are all originally put into an STM class.
For simplicity I have renamed this class to LKD and separated out this
change to highlight modification outside of the lkd/ directory.

My assumption is that the specific commands could/should be added to
other classes - but I've left things as they are for now.