lkd-main: Provide !CONFIG_MMU compatible symbol check
Some symbol files, that ordinarily you might expect to find, are not
available on MMU-Less builds. To support this seamlessly, we can adjust
the symbols we check when we try to identify if we have debug symbols.
We still perform the existing checks to ensure that any expected
behaviour stays the same