DNI: Investigation into gdb.Value string bug
So far, it looks like there is some sort of an object cache, and objects
are re-used. When a user performs a 'bt' operation, it is reading through
the dwarf2-frame functions and something in here seems to be (re?-)initialising
the gdb.Value which represents the string for the lx-version string.

Interesting to note is that it only occurs on the lx-version string, which
is presumably being read from the vmlinux binary rather than memory...?